adds cli to __main__
