+++ bootstrap/__main__.py
@@ -1,0 +1,7 @@
+__author__ = "Bernhard Esperester <bernhard@esperester.de>"
+
+from bootstrap.cli import main
+
+
+if __name__ == "__main__":
+    main()
